[Sparse] Remove the Python DiagMatrix class
