+++ python/dgl/sparse/matmul.py
@@ -4,19 +4,17 @@ from typing import Union
 
 import torch
 
-from .diag_matrix import diag, DiagMatrix
-
-from .sparse_matrix import SparseMatrix, val_like
+from .sparse_matrix import SparseMatrix
 
 __all__ = ["spmm", "bspmm", "spspmm", "matmul"]
 
 
-def spmm(A: Union[SparseMatrix, DiagMatrix], X: torch.Tensor) -> torch.Tensor:
+def spmm(A: SparseMatrix, X: torch.Tensor) -> torch.Tensor:
     """Multiplies a sparse matrix by a dense matrix, equivalent to ``A @ X``.
 
     Parameters
     ----------
-    A : SparseMatrix or DiagMatrix
+    A : SparseMatrix
         Sparse matrix of shape ``(L, M)`` with scalar values
     X : torch.Tensor
         Dense matrix of shape ``(M, N)`` or ``(M)``
@@ -30,7 +28,7 @@ def spmm(A: Union[SparseMatrix, DiagMatrix], X: torch.Tensor) -> torch.Tensor:
     --------
 
     >>> indices = torch.tensor([[0, 1, 1], [1, 0, 1]])
-    >>> val = torch.randn(len(row))
+    >>> val = torch.randn(indices.shape[1])
     >>> A = dglsp.spmatrix(indices, val)
     >>> X = torch.randn(2, 3)
     >>> result = dglsp.spmm(A, X)
@@ -40,25 +38,22 @@ def spmm(A: Union[SparseMatrix, DiagMatrix], X: torch.Tensor) -> torch.Tensor:
     torch.Size([2, 3])
     """
     assert isinstance(
-        A, (SparseMatrix, DiagMatrix)
-    ), f"Expect arg1 to be a SparseMatrix or DiagMatrix object, got {type(A)}."
+        A, SparseMatrix
+    ), f"Expect arg1 to be a SparseMatrix object, got {type(A)}."
     assert isinstance(
         X, torch.Tensor
     ), f"Expect arg2 to be a torch.Tensor, got {type(X)}."
 
-    # The input is a DiagMatrix. Cast it to SparseMatrix
-    if not isinstance(A, SparseMatrix):
-        A = A.to_sparse()
     return torch.ops.dgl_sparse.spmm(A.c_sparse_matrix, X)
 
 
-def bspmm(A: Union[SparseMatrix, DiagMatrix], X: torch.Tensor) -> torch.Tensor:
+def bspmm(A: SparseMatrix, X: torch.Tensor) -> torch.Tensor:
     """Multiplies a sparse matrix by a dense matrix by batches, equivalent to
     ``A @ X``.
 
     Parameters
     ----------
-    A : SparseMatrix or DiagMatrix
+    A : SparseMatrix
         Sparse matrix of shape ``(L, M)`` with vector values of length ``K``
     X : torch.Tensor
         Dense matrix of shape ``(M, N, K)``
@@ -82,115 +77,30 @@ def bspmm(A: Union[SparseMatrix, DiagMatrix], X: torch.Tensor) -> torch.Tensor:
     torch.Size([3, 3, 2])
     """
     assert isinstance(
-        A, (SparseMatrix, DiagMatrix)
-    ), f"Expect arg1 to be a SparseMatrix or DiagMatrix object, got {type(A)}."
+        A, SparseMatrix
+    ), f"Expect arg1 to be a SparseMatrix object, got {type(A)}."
     assert isinstance(
         X, torch.Tensor
     ), f"Expect arg2 to be a torch.Tensor, got {type(X)}."
     return spmm(A, X)
 
 
-def _diag_diag_mm(A: DiagMatrix, B: DiagMatrix) -> DiagMatrix:
-    """Internal function for multiplying a diagonal matrix by a diagonal matrix.
-
-    Parameters
-    ----------
-    A : DiagMatrix
-        Diagonal matrix of shape ``(L, M)``
-    B : DiagMatrix
-        Diagonal matrix of shape ``(M, N)``
-
-    Returns
-    -------
-    DiagMatrix
-        Diagonal matrix of shape ``(L, N)``
-    """
-    M, N = A.shape
-    N, P = B.shape
-    common_diag_len = min(M, N, P)
-    new_diag_len = min(M, P)
-    diag_val = torch.zeros(new_diag_len)
-    diag_val[:common_diag_len] = (
-        A.val[:common_diag_len] * B.val[:common_diag_len]
-    )
-    return diag(diag_val.to(A.device), (M, P))
-
-
-def _sparse_diag_mm(A, D):
-    """Internal function for multiplying a sparse matrix by a diagonal matrix.
-
-    Parameters
-    ----------
-    A : SparseMatrix
-        Sparse matrix of shape ``(L, M)``
-    D : DiagMatrix
-        Diagonal matrix of shape ``(M, N)``
-
-    Returns
-    -------
-    SparseMatrix
-        Sparse matrix of shape ``(L, N)``
-    """
-    assert (
-        A.shape[1] == D.shape[0]
-    ), f"The second dimension of SparseMatrix should be equal to the first \
-    dimension of DiagMatrix in matmul(SparseMatrix, DiagMatrix), but the \
-    shapes of SparseMatrix and DiagMatrix are {A.shape} and {D.shape} \
-    respectively."
-    assert (
-        D.shape[0] == D.shape[1]
-    ), f"The DiagMatrix should be a square in matmul(SparseMatrix, DiagMatrix) \
-    but got {D.shape}."
-    return val_like(A, D.val[A.col] * A.val)
-
-
-def _diag_sparse_mm(D, A):
-    """Internal function for multiplying a diagonal matrix by a sparse matrix.
-
-    Parameters
-    ----------
-    D : DiagMatrix
-        Diagonal matrix of shape ``(L, M)``
-    A : SparseMatrix
-        Sparse matrix of shape ``(M, N)``
-
-    Returns
-    -------
-    SparseMatrix
-        Sparse matrix of shape ``(L, N)``
-    """
-    assert (
-        D.shape[1] == A.shape[0]
-    ), f"The second dimension of DiagMatrix should be equal to the first \
-    dimension of SparseMatrix in matmul(DiagMatrix, SparseMatrix), but the \
-    shapes of DiagMatrix and SparseMatrix are {D.shape} and {A.shape} \
-    respectively."
-    assert (
-        D.shape[0] == D.shape[1]
-    ), f"The DiagMatrix should be a square in matmul(DiagMatrix, SparseMatrix) \
-    but got {D.shape}."
-    return val_like(A, D.val[A.row] * A.val)
-
-
-def spspmm(
-    A: Union[SparseMatrix, DiagMatrix], B: Union[SparseMatrix, DiagMatrix]
-) -> Union[SparseMatrix, DiagMatrix]:
+def spspmm(A: SparseMatrix, B: SparseMatrix) -> SparseMatrix:
     """Multiplies a sparse matrix by a sparse matrix, equivalent to ``A @ B``.
 
     The non-zero values of the two sparse matrices must be 1D.
 
     Parameters
     ----------
-    A : SparseMatrix or DiagMatrix
+    A : SparseMatrix
         Sparse matrix of shape ``(L, M)``
-    B : SparseMatrix or DiagMatrix
+    B : SparseMatrix
         Sparse matrix of shape ``(M, N)``
 
     Returns
     -------
-    SparseMatrix or DiagMatrix
-        Matrix of shape ``(L, N)``. It is a DiagMatrix object if both matrices
-        are DiagMatrix objects, otherwise a SparseMatrix object.
+    SparseMatrix
+        Sparse matrix of shape ``(L, N)``.
 
     Examples
     --------
@@ -208,68 +118,52 @@ def spspmm(
                  shape=(2, 3), nnz=5)
     """
     assert isinstance(
-        A, (SparseMatrix, DiagMatrix)
-    ), f"Expect A1 to be a SparseMatrix or DiagMatrix object, got {type(A)}."
+        A, SparseMatrix
+    ), f"Expect A1 to be a SparseMatrix object, got {type(A)}."
     assert isinstance(
-        B, (SparseMatrix, DiagMatrix)
-    ), f"Expect A2 to be a SparseMatrix or DiagMatrix object, got {type(B)}."
+        B, SparseMatrix
+    ), f"Expect A2 to be a SparseMatrix object, got {type(B)}."
 
-    if isinstance(A, DiagMatrix) and isinstance(B, DiagMatrix):
-        return _diag_diag_mm(A, B)
-    if isinstance(A, DiagMatrix):
-        return _diag_sparse_mm(A, B)
-    if isinstance(B, DiagMatrix):
-        return _sparse_diag_mm(A, B)
     return SparseMatrix(
         torch.ops.dgl_sparse.spspmm(A.c_sparse_matrix, B.c_sparse_matrix)
     )
 
 
 def matmul(
-    A: Union[torch.Tensor, SparseMatrix, DiagMatrix],
-    B: Union[torch.Tensor, SparseMatrix, DiagMatrix],
-) -> Union[torch.Tensor, SparseMatrix, DiagMatrix]:
-    """Multiplies two dense/sparse/diagonal matrices, equivalent to ``A @ B``.
+    A: Union[torch.Tensor, SparseMatrix], B: Union[torch.Tensor, SparseMatrix]
+) -> Union[torch.Tensor, SparseMatrix]:
+    """Multiplies two dense/sparse matrices, equivalent to ``A @ B``.
 
-    The supported combinations are shown as follows.
-
-    +--------------+--------+------------+--------------+
-    |   A \\ B   | Tensor | DiagMatrix | SparseMatrix |
-    +--------------+--------+------------+--------------+
-    |    Tensor    |   ✅   |     🚫     |      🚫      |
-    +--------------+--------+------------+--------------+
-    | SparseMatrix |   ✅   |     ✅     |      ✅      |
-    +--------------+--------+------------+--------------+
-    |  DiagMatrix  |   ✅   |     ✅     |      ✅      |
-    +--------------+--------+------------+--------------+
+    This function does not support the case where :attr:`A` is a \
+    ``torch.Tensor`` and :attr:`B` is a ``SparseMatrix``.
 
     * If both matrices are torch.Tensor, it calls \
         :func:`torch.matmul()`. The result is a dense matrix.
 
-    * If both matrices are sparse or diagonal, it calls \
-        :func:`dgl.sparse.spspmm`. The result is a sparse matrix.
+    * If both matrices are sparse, it calls :func:`dgl.sparse.spspmm`. The \
+        result is a sparse matrix.
 
-    * If :attr:`A` is sparse or diagonal while :attr:`B` is dense, it \
-        calls :func:`dgl.sparse.spmm`. The result is a dense matrix.
+    * If :attr:`A` is sparse while :attr:`B` is dense, it calls \
+        :func:`dgl.sparse.spmm`. The result is a dense matrix.
 
     * The operator supports batched sparse-dense matrix multiplication. In \
-        this case, the sparse or diagonal matrix :attr:`A` should have shape \
-        ``(L, M)``, where the non-zero values have a batch dimension ``K``. \
-        The dense matrix :attr:`B` should have shape ``(M, N, K)``. The output \
+        this case, the sparse matrix :attr:`A` should have shape ``(L, M)``, \
+        where the non-zero values have a batch dimension ``K``. The dense \
+        matrix :attr:`B` should have shape ``(M, N, K)``. The output \
         is a dense matrix of shape ``(L, N, K)``.
 
     * Sparse-sparse matrix multiplication does not support batched computation.
 
     Parameters
     ----------
-    A : torch.Tensor, SparseMatrix or DiagMatrix
+    A : torch.Tensor or SparseMatrix
         The first matrix.
-    B : torch.Tensor, SparseMatrix, or DiagMatrix
+    B : torch.Tensor or SparseMatrix
         The second matrix.
 
     Returns
     -------
-    torch.Tensor, SparseMatrix or DiagMatrix
+    torch.Tensor or SparseMatrix
         The result matrix
 
     Examples
@@ -289,7 +183,7 @@ def matmul(
     Multiplies a sparse matrix with a dense matrix.
 
     >>> indices = torch.tensor([[0, 1, 1], [1, 0, 1]])
-    >>> val = torch.randn(len(row))
+    >>> val = torch.randn(indices.shape[1])
     >>> A = dglsp.spmatrix(indices, val)
     >>> X = torch.randn(2, 3)
     >>> result = dglsp.matmul(A, X)
@@ -301,10 +195,10 @@ def matmul(
     Multiplies a sparse matrix with a sparse matrix.
 
     >>> indices1 = torch.tensor([[0, 1, 1], [1, 0, 1]])
-    >>> val1 = torch.ones(len(row1))
+    >>> val1 = torch.ones(indices1.shape[1])
     >>> A = dglsp.spmatrix(indices1, val1)
     >>> indices2 = torch.tensor([[0, 1, 1], [0, 2, 1]])
-    >>> val2 = torch.ones(len(row2))
+    >>> val2 = torch.ones(indices2.shape[1])
     >>> B = dglsp.spmatrix(indices2, val2)
     >>> result = dglsp.matmul(A, B)
     >>> type(result)
@@ -312,12 +206,11 @@ def matmul(
     >>> result.shape
     (2, 3)
     """
-    assert isinstance(A, (torch.Tensor, SparseMatrix, DiagMatrix)), (
-        f"Expect arg1 to be a torch.Tensor, SparseMatrix, or DiagMatrix object,"
-        f"got {type(A)}."
-    )
-    assert isinstance(B, (torch.Tensor, SparseMatrix, DiagMatrix)), (
-        f"Expect arg2 to be a torch Tensor, SparseMatrix, or DiagMatrix"
+    assert isinstance(
+        A, (torch.Tensor, SparseMatrix)
+    ), f"Expect arg1 to be a torch.Tensor or SparseMatrix, got {type(A)}."
+    assert isinstance(B, (torch.Tensor, SparseMatrix)), (
+        f"Expect arg2 to be a torch Tensor or SparseMatrix"
         f"object, got {type(B)}."
     )
     if isinstance(A, torch.Tensor) and isinstance(B, torch.Tensor):
@@ -328,10 +221,7 @@ def matmul(
     )
     if isinstance(B, torch.Tensor):
         return spmm(A, B)
-    if isinstance(A, DiagMatrix) and isinstance(B, DiagMatrix):
-        return _diag_diag_mm(A, B)
     return spspmm(A, B)
 
 
 SparseMatrix.__matmul__ = matmul
-DiagMatrix.__matmul__ = matmul
