+++ tests/python/pytorch/sparse/test_elementwise_op.py
@@ -1,5 +1,4 @@
 import operator
-import sys
 
 import backend as F
